backend/requirements.txt
flask
sqlalchemy
alembic
bcrypt
PyJWT
pytest
pytest-cov
apscheduler
psycopg2-binary
flask-cors
pywebpush
